L0: Replaced Space Cadet shifts with normal shifts. Swapped AltGr/App, space, and Enter. L3: added Teensy reset on ESC key.
I have a ton of trouble with the spacebar being so far away. I'm going to try moving space and Return over once just to see if that's doable, as I realize that on normal keyboards I keep my right thumb very close to the M key.

I also have trouble with the space cadet shifts. There are plenty of times when I type way too quickly for the Teensy board to pick up the shift and not the parenthesis. I'll turn it off and try to focus on using L1 for now.

Finally, I would like to be able to access the Teensy bootloader without finding a damn paperclip.